src/uma_calculator.py
"""
Stage 2: UMA Energy Calculation

This module uses Meta's UMA model to compute the ground-state reaction energy
for the hydrolysis of PLA oligomers.
"""

from ase.optimize import LBFGS
from fairchem.core import pretrained_mlip, FAIRChemCalculator
from ase.build import molecule
import numpy as np

def setup_calculator(model_name="uma-s-1p2", device="cpu"):
    """
    Initialize the UMA calculator.
    """
    predictor = pretrained_mlip.get_predict_unit(model_name, device=device)
    calc = FAIRChemCalculator(predictor, task_name="omol")
    return calc

def calculate_hydrolysis_reaction_energy(oligomer_atoms, calc):
    """
    Calculate the reaction energy for ester hydrolysis:
      PLA_n + H2O  ->  PLA_(n-1) + lactic acid
    
    Returns ΔE_rxn in eV.
    """
    # Note: This is a placeholder implementation.
    # In a full implementation, you would carefully construct the reactant
    # complex (oligomer + water) and the product complex (cleaved chains),
    # relax both, and take the energy difference.
    
    # Example for relaxing a single structure:
    # oligomer_atoms.calc = calc
    # opt = LBFGS(oligomer_atoms, logfile=None)
    # opt.run(fmax=0.05, steps=200)
    # energy = oligomer_atoms.get_potential_energy()
    
    return -0.35  # Placeholder dummy value

if __name__ == "__main__":
    print("UMA calculator module ready.")
