requirements.txt
fairchem-core>=2.0.0
ase>=3.22.1
rdkit-pypi>=2022.9.5
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
streamlit>=1.30.0
huggingface_hub>=0.20.0
jupyter>=1.0.0
